Split the basic issue table and issue table filter tests
